fix long result array

+++ main.cu
@@ -40,7 +40,7 @@ __global__ void cu_gen_distance_array(int *distance_array_d, int max) {
 extern "C" void gpu_dotProduct(long long *result_array, int num_vectors) {
   int *distance_array_d;
   int *force_array_d;
-  int *result_array_d;
+  long long *result_array_d;
 
   // allocate space in the device
   cudaMalloc((void **)&distance_array_d, sizeof(int) * num_vectors);
